Added more Bus and Monkey logic
I've decided that the bus will control it's position but the map matrix will pull it every so often to check for collisions and out of bounds cases.

We currently lack a way to manage the controls and what controls would be available (e.g fading an option when it is not viable).